requirements.txt
boto3
discord.py
python-dotenv
pytz
pandas
